Begin to update the python API to reflect collections.
This commit reflects all of the playlist operations being updated to include a playlist option. It also removes the medialib methods that were made part of the playlist, but does not add them to the playlist. It also adds a from_list method which is supposed to convert a python list to a char ** type array, but this function does not currently work because pyrex refuses to return a non python object:
/usr/home/alex/xmms2-alex/src/clients/lib/python/xmmsclient.pyx:251:8: Cannot convert 'char (*(*))' to Python object