Added transactions to the bdb code
Made the BDB-code ACID-compliant using transactions... And it turned out
to be slow! Even when using DB_TXN_NOSYNC it takes a LONG time (I don't
really know how long, after 4 minutes I killed it). Most of this is
because of disk latency (I do have a shitty laptop disk, so I guess it
would be faster on a desktop with 7.2k disks). There might be something
I'm doing wrong, but it seems transactions are just too slow.