Locking and syncing for the homegrown backend
Added a Multiple Readers, Single Writer lock to the database, that
should make it thread safe. Also added some sync routines. Whenever
someone grabs a write lock the database is marked as dirty. The sync
thread runs every 60 second and checks if the database is dirty, if it
is it syncs it.