B+ tree
Added a B+ tree to see if that would be any faster than a linked list
for the int-pair store. It seems to be quite a bit faster, and with some
tweaking of the node size I don't think it will take that much more
space. The code is in need of some major cleanup though.