B+ tree working (it seems like)
Fixed the last bug and rewrote the search to use binary search. Didn't
do much difference, but it got a little bit faster. Some time
comparisons with the old linked list:
|    B+ tree  |     List    |
| user | sys  | user | sys  |
| ~0.9 | ~1.0 | ~5.3 | ~2.9 |
So, it's quite a bit faster, and it doesn't use that much more space
(14 MB vs. 12 MB with 63 keys/node).