FEATURE(2262): WIP - Working trie-based completion of commands and subcommands. Still ini the works, I'm just pushing this to share the fun, but code will still be improved!
It notably works in auto-complete mode and with aliases.
All of this only works in readline/shell-mode, obviously.
There is no completion for arguments yet.

The trie complete function have been duplicated from the find
functions; in the end, the same functions should probably be used to
perform both the matching of actions and generation of completions.

Ah and I need to write tests for this stuff, seriously.